internalized mission status updated and added errors throwing

+++ src/Controller/MissionController.php
@@ -49,6 +49,11 @@ class MissionController extends Controller
 
         $missions = $repository->findBy(array(), array('id' => 'DESC'));
 
+        if (!$missions) {
+            throw $this->createNotFoundException('Aucune fiche mission trouvée.');
+
+        }
+
         return $this->render('mission/list.html.twig', array(
             'missions' => $missions
         ));
@@ -118,6 +123,11 @@ class MissionController extends Controller
     */
     public function view(Mission $mission)
     {
+        if (!$mission) {
+            throw $this->createNotFoundException('La fiche mission demandée n\'existe pas.');
+
+        }
+
         return $this->render('mission/view.html.twig', array(
             'mission' => $mission
         ));
@@ -136,6 +146,11 @@ class MissionController extends Controller
     */
     public function edit(Mission $mission, Request $request, FileUploader $fileUploader)
     {
+        if (!$mission) {
+            throw $this->createNotFoundException('La fiche mission demandée n\'existe pas.');
+
+        }
+
         // Saving previous scan to avoid delete by edit
         $previousFileName = $mission->getAttachment();
 
@@ -151,17 +166,18 @@ class MissionController extends Controller
         if ($form->isSubmitted() && $form->isValid()) {
             
             /****************** Update status ******************/
-            $status = $mission->getStatus(); // Current mission status
+            // $status = $mission->getStatus(); // Current mission status
 
-            if ($mission->getDateFinished() !== null && $status !== Mission::STATUS_FINISHED) {
-                $mission->setStatus(Mission::STATUS_FINISHED);
-            }
-            else if ($mission->getDateAssigned() !== null && $status !== Mission::STATUS_ASSIGNED) {
-                $mission->setStatus(Mission::STATUS_ASSIGNED);
-            }
-            else if ($status !== Mission::STATUS_DEFAULT) {
-                $mission->setStatus(Mission::STATUS_DEFAULT);  // [shouldn't happen] Set to STATUS_DEFAULT
-            }
+            // if ($mission->getDateFinished() !== null && $status !== Mission::STATUS_FINISHED) {
+            //     $mission->setStatus(Mission::STATUS_FINISHED);
+            // }
+            // else if ($mission->getDateAssigned() !== null && $status !== Mission::STATUS_ASSIGNED) {
+            //     $mission->setStatus(Mission::STATUS_ASSIGNED);
+            // }
+            // else if ($status !== Mission::STATUS_DEFAULT) {
+            //     $mission->setStatus(Mission::STATUS_DEFAULT);  // Shouldn't happen
+            // }
+            $mission->updateStatus();
 
             /**************** Manage attachment ****************/
             /** @var Symfony\Component\HttpFoundation\File\UploadedFile $file */
@@ -209,6 +225,11 @@ class MissionController extends Controller
     */
     public function delete(Mission $mission, FileUploader $fileUploader)
     {
+        if (!$mission) {
+            throw $this->createNotFoundException('La fiche mission demandée n\'existe pas.');
+
+        }
+
         // Delete attached file from server
         $fileName = $mission->getAttachment();
         $fileUploader->delete($fileName);
@@ -224,130 +245,6 @@ class MissionController extends Controller
 
         return $this->redirectToRoute('app_missions_list');
     }
-
-    /**
-    * Recap of opened missions
-    *
-    * @Route(
-    *  "/recap/{page}",
-    *  name="app_missions_recap",
-    *  requirements={
-    *      "page"="\d+"
-    *  }
-    * )
-    */
-    public function recap($page = 1)
-    {    
-        $missions = $this->getNonClosedMissions();
-
-        return $this->render('mission/recap.html.twig', array(
-            'missions' => $missions
-        ));
-    }
-
-    /**
-    * Empty field attachment and delete file
-    * 
-    * @Route(
-    *  "/delete-attachment/{id}",
-    *  name="app_missions_attachment-delete",
-    *  requirements={
-    *      "id"="\d+"
-    *  }
-    * )
-    */
-    public function deleteAttachment(Mission $mission, FileUploader $fileUploader)
-    { 
-        // Delete file from server
-        $fileName = $mission->getAttachment();
-
-        $fileUploader->delete($fileName);
-
-        // Empty field in database
-        $mission->setAttachment(null);
-
-        $em = $this->getDoctrine()->getManager();
-        $em->persist($mission);
-        $em->flush();
-
-        return $this->redirectToRoute('app_missions_view', array(
-                'id' => $mission->getId()
-        ));
-    }
-
-    /**
-     * Export a given mission to PDF
-     * 
-     * @Route(
-     *  "/pdf/{id}",
-     *  name="app_missions_pdf-export",
-     *  requirements={
-     *      "id"="\d+"
-     *  }
-     * )
-     */
-    public function pdfExport(Mission $mission)
-    {
-        $html = $this->renderView('pdf/mission.html.twig', array(
-            'mission' => $mission
-        ));
-
-        $filename = sprintf('fm-%s_%s.pdf', $mission->getId(), date('Y-m-d_His'));
-
-        return new Response(
-            $this->get('knp_snappy.pdf')->getOutputFromHtml($html),
-            200,
-            [
-                'Content-Type'        => 'application/pdf',
-                'Content-Disposition' => sprintf('attachment; filename="%s"', $filename),
-            ]
-        );
-    }
-
-    /**
-     * Export recap of opened missions to PDF
-     * 
-     * @Route(
-     *  "/recap/pdf",
-     *  name="app_missions_recap_pdf-export"
-     * )
-     */
-    public function recapPdfExport()
-    {
-        $missions = $this->getNonClosedMissions();
-
-        $html = $this->renderView('pdf/missions-recap.html.twig', array(
-            'missions' => $missions
-        ));
-
-        $filename = sprintf('fm-recap_%s.pdf', date('Y-m-d_His'));
-
-        return new Response(
-            $this->get('knp_snappy.pdf')->getOutputFromHtml($html, array(
-                'page-size' => 'A3',
-                'orientation' => 'Landscape'
-            )),
-            200,
-            [
-                'Content-Type'        => 'application/pdf',
-                'Content-Disposition' => sprintf('attachment; filename="%s"', $filename),
-            ]
-        );
-    }
-
-    private function getNonClosedMissions()
-    {
-        $statuses = array_values(Mission::getStatuses()); // List of possible statuses
-        array_pop($statuses); // Remove STATUS_CLOSED
-
-        $repository = $this->getDoctrine()->getRepository(Mission::class);
-
-        // Get all missions with status not "Fermée"
-        $missions = $repository->findByStatus($statuses);
-
-        return $missions;
-    }
-
 
     /**
      * Close / re-open a mission if its current status is finished / closed respectively
@@ -395,4 +292,130 @@ class MissionController extends Controller
             'id' => $mission->getId()
         ));
     }
+
+
+        /**
+    * Empty field "attachment" and delete file
+    * 
+    * @Route(
+    *  "/delete-attachment/{id}",
+    *  name="app_missions_attachment-delete",
+    *  requirements={
+    *      "id"="\d+"
+    *  }
+    * )
+    */
+    public function deleteAttachment(Mission $mission, FileUploader $fileUploader)
+    { 
+        // Delete file from server
+        $fileName = $mission->getAttachment();
+        $fileUploader->delete($fileName);
+
+        // Empty field in database
+        $mission->setAttachment(null);
+
+        $em = $this->getDoctrine()->getManager();
+        $em->persist($mission);
+        $em->flush();
+
+        return $this->redirectToRoute('app_missions_view', array(
+                'id' => $mission->getId()
+        ));
+    }
+
+    /**
+     * Export a given mission to PDF
+     * 
+     * @Route(
+     *  "/pdf/{id}",
+     *  name="app_missions_pdf-export",
+     *  requirements={
+     *      "id"="\d+"
+     *  }
+     * )
+     */
+    public function pdfExport(Mission $mission)
+    {
+        $html = $this->renderView('pdf/mission.html.twig', array(
+            'mission' => $mission
+        ));
+
+        $filename = sprintf('fm-%s_%s.pdf', $mission->getId(), date('Y-m-d_His'));
+
+        return new Response(
+            $this->get('knp_snappy.pdf')->getOutputFromHtml($html),
+            200,
+            [
+                'Content-Type'        => 'application/pdf',
+                'Content-Disposition' => sprintf('attachment; filename="%s"', $filename),
+            ]
+        );
+    }
+
+
+    /**
+    * Recap of opened (= any status but "closed") missions
+    *
+    * @Route(
+    *  "/recap/{page}",
+    *  name="app_missions_recap",
+    *  requirements={
+    *      "page"="\d+"
+    *  }
+    * )
+    */
+    public function recap($page = 1)
+    {    
+        $missions = $this->getNonClosedMissions();
+
+        return $this->render('mission/recap.html.twig', array(
+            'missions' => $missions
+        ));
+    }
+
+
+    /**
+     * Export recap of opened missions to PDF
+     * 
+     * @Route(
+     *  "/recap/pdf",
+     *  name="app_missions_recap_pdf-export"
+     * )
+     */
+    public function recapPdfExport()
+    {
+        $missions = $this->getNonClosedMissions();
+
+        $html = $this->renderView('pdf/missions-recap.html.twig', array(
+            'missions' => $missions
+        ));
+
+        $filename = sprintf('fm-recap_%s.pdf', date('Y-m-d_His'));
+
+        return new Response(
+            $this->get('knp_snappy.pdf')->getOutputFromHtml($html, array(
+                'page-size' => 'A3',
+                'orientation' => 'Landscape'
+            )),
+            200,
+            [
+                'Content-Type'        => 'application/pdf',
+                'Content-Disposition' => sprintf('attachment; filename="%s"', $filename),
+            ]
+        );
+    }
+
+    private function getNonClosedMissions()
+    {
+        $statuses = array_values(Mission::getStatuses()); // Array of possible statuses
+        array_pop($statuses); // Remove STATUS_CLOSED
+
+        $repository = $this->getDoctrine()->getRepository(Mission::class);
+
+        // Get all missions with status not "Fermée"
+        $missions = $repository->findByStatus($statuses);
+
+        return $missions;
+    }
+
 }
